Renamed dirs and removed unrequired UI buttons
