+++ data/com.adhadse.Bookx.gresource.xml
@@ -3,12 +3,13 @@
   <gresource prefix="/com/adhadse/Bookx/">
     <!-- see https://gtk-rs.org/gtk4-rs/git/docs/gtk4/struct.Application.html#automatic-resources -->
     <file compressed="true" preprocess="xml-stripblanks" alias="gtk/help-overlay.ui">ui/shortcuts.ui</file>
-    <file compressed="true" preprocess="xml-stripblanks">ui/window.ui</file>
     <file compressed="true" preprocess="xml-stripblanks">ui/library.ui</file>
+    <file compressed="true" preprocess="xml-stripblanks">ui/library_carousal.ui</file>
+    <file compressed="true" preprocess="xml-stripblanks">ui/library_carousal_component.ui</file>
+    <file compressed="true" preprocess="xml-stripblanks">ui/library_content_box.ui</file>
+    <file compressed="true" preprocess="xml-stripblanks">ui/window.ui</file>
     <file compressed="true" preprocess="xml-stripblanks">ui/preferences.ui</file>
     <file compressed="true" preprocess="xml-stripblanks">ui/shortcuts.ui</file>
-    <file compressed="true" preprocess="xml-stripblanks">ui/subpage_library_carousal_page.ui</file>
-    <file compressed="true" preprocess="xml-stripblanks">ui/subpage_library_content_page.ui</file>
     <file compressed="true">style.css</file>
   </gresource>
 </gresources>
