+++ data/ui/library_carousal_component.ui
