+++ src/ui/mod.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-pub mod pages;
+pub mod library;
 
 mod window;
 pub use window::BookxView;
